requirements.txt
asgiref==3.9.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-humanize==0.1.2
humanize==4.13.0
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
